add default identity user and JWT Auth
Identity user is ignoring the custom user we have, but currently we have no reason to use any custom fields on our user, so we should remove the custom user prior to merging. 'Scope' is handled via 'Roles' with .NET identity. This change does not have any way to issue JWT or authenticate using Google Oauth, but should be a good starting point for us to implement those prior to merging.