remove user roles handling, JWT auth and acc creation is working
I could not find any way to create new roles or use the built in role handler, but some of the boiler plate is here. I removed our custom user model in favor of the built-in identity user. All endpoints for creating, deleting, or manipulating recipe data should require jwt auth now. Any user can create an account though, so the endpoints are not secured until we implement user roles.

There is a lot of manual work with the JWTs here and I suspect more is meant to be built-in to the framework, but maybe not as I could not find documentation for it.