Add Users and JWT auth (#9)
* add user model, add google oauth config"

* it builds, but unsure if it works, think this method is how to do it

* remove jwt from response for testing

* add default identity user and JWT Auth

Identity user is ignoring the custom user we have, but currently we have no reason to use any custom fields on our user, so we should remove the custom user prior to merging. 'Scope' is handled via 'Roles' with .NET identity. This change does not have any way to issue JWT or authenticate using Google Oauth, but should be a good starting point for us to implement those prior to merging.

* Remove custom JWT generation

This JWT script is no longer needed as we can use .NET built-in claims + identity to generate and validate JWT

* remove reference to removed JWT generator

* add endpoint to login a user and issue a JWT

This bypasses a lot of the .AddJwtBearer options. This should be changed in the future to better utilize the built-ins that I assume should be handling this with less manual configuration in the controller.

* add user registration and swagger docs for JWT auth

* broken commit, attempt to add user roles

It appears there are no default roles included with the built-in role manager, or the defaults are not documented. There is no documentation on creating your own roles, so we will need to figure this out.

* remove user roles handling, JWT auth and acc creation is working

I could not find any way to create new roles or use the built in role handler, but some of the boiler plate is here. I removed our custom user model in favor of the built-in identity user. All endpoints for creating, deleting, or manipulating recipe data should require jwt auth now. Any user can create an account though, so the endpoints are not secured until we implement user roles.

There is a lot of manual work with the JWTs here and I suspect more is meant to be built-in to the framework, but maybe not as I could not find documentation for it.

* add required environment vars to launchSettings.json

* update readme to include env var setup

* remove env var validation from database context class

* update migrations

* split up auth and user endpoints, create DTOs, refactor responses

* return username instead of user object at user creation

* fix incorrect indentation

* add missing env checks to startup

* remove unreachable return

* replace BadRequest with Unauthorized for invalid login

* add profile fetch from JWT

* change JWT to store user ID and use user ID to lookup user for profile